Refactor columnWrapper
